# docs/openapi-gl.yaml
openapi: 3.0.3
info:
  title: Smart Accounting - General Ledger API
  version: 1.0.0
  description: |
    GL Module API compliant with Circular 99/2025/TT-BTC
    and Accounting Law 88/2015/QH13.

servers:
  - url: http://localhost:8080/api/gl

paths:
  # ===== ACCOUNTS =====
  /accounts:
    post:
      summary: Create new account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      summary: List all accounts
      tags: [Accounts]
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'

  /accounts/tree:
    get:
      summary: Get account hierarchy tree
      tags: [Accounts]
      responses:
        '200':
          description: Account tree
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountTreeNode'

  /accounts/active:
    get:
      summary: List active accounts only
      tags: [Accounts]
      responses:
        '200':
          description: Active accounts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'

  /accounts/{id}:
    get:
      summary: Get account by ID
      tags: [Accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update account
      tags: [Accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                nameEn:
                  type: string
                description:
                  type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Account updated
    delete:
      summary: Deactivate account (soft delete)
      tags: [Accounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Account deactivated

  /accounts/import:
    post:
      summary: Bulk import chart of accounts
      tags: [Accounts]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accounts:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Accounts imported

  # ===== JOURNALS =====
  /journals:
    post:
      summary: Create journal batch
      tags: [Journals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalRequest'
      responses:
        '201':
          description: Journal created
    get:
      summary: Search journal batches
      tags: [Journals]
      parameters:
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JournalEntryStatus'
        - name: accountId
          in: query
          schema:
            type: string
        - name: periodId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Journal list

  /journals/{id}:
    get:
      summary: Get journal by ID
      tags: [Journals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Journal details
    delete:
      summary: Delete journal (draft only)
      tags: [Journals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted

  /journals/{id}/submit:
    post:
      summary: Submit journal for approval
      tags: [Journals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Submitted

  /journals/{id}/approve:
    post:
      summary: Approve journal
      tags: [Journals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved

  /journals/{id}/post:
    post:
      summary: Post journal (update account balances)
      tags: [Journals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Posted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostingResult'

  /journals/{id}/reverse:
    post:
      summary: Reverse a posted journal
      tags: [Journals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
      responses:
        '200':
          description: Reversed

  /journals/validate:
    post:
      summary: Validate a batch without posting
      tags: [Journals]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                batchId:
                  type: string
      responses:
        '200':
          description: Validation result

  /journals/batch-post:
    post:
      summary: Post multiple batches
      tags: [Journals]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                batchIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Batch posting results

  # ===== FISCAL YEARS & PERIODS =====
  /periods/fiscal-years:
    post:
      summary: Create fiscal year with periods
      tags: [Periods]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFiscalYearRequest'
      responses:
        '201':
          description: Fiscal year created
    get:
      summary: List fiscal years
      tags: [Periods]
      responses:
        '200':
          description: Fiscal year list

  /periods/fiscal-years/active:
    get:
      summary: Get active fiscal year
      tags: [Periods]
      responses:
        '200':
          description: Active fiscal year

  /periods/fiscal-years/{id}:
    get:
      summary: Get fiscal year by ID
      tags: [Periods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fiscal year details

  /periods/fiscal-years/{id}/close:
    post:
      summary: Close fiscal year
      tags: [Periods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fiscal year closed

  /periods/fiscal-years/{id}/periods:
    get:
      summary: Get periods for fiscal year
      tags: [Periods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Period list

  /periods/periods/{id}/close:
    post:
      summary: Close period
      tags: [Periods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Period closed

  /periods/periods/{id}/reopen:
    post:
      summary: Reopen period
      tags: [Periods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Period reopened

  /periods/periods/{id}/lock:
    post:
      summary: Lock period (prevent any posting)
      tags: [Periods]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Period locked

components:
  schemas:
    CreateAccountRequest:
      type: object
      required: [code, name, category, nature]
      properties:
        code:
          type: string
          example: "1111"
        name:
          type: string
          example: "Tiền mặt VND"
        category:
          $ref: '#/components/schemas/AccountCategory'
        nature:
          $ref: '#/components/schemas/AccountNature'
        parentId:
          type: string
        currencyCode:
          type: string
          example: VND
        description:
          type: string
        isControl:
          type: boolean
        isPosting:
          type: boolean

    AccountResponse:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        category:
          type: string
        nature:
          type: string
        balance:
          type: number
        isActive:
          type: boolean
        isControl:
          type: boolean
        isPosting:
          type: boolean
        version:
          type: integer

    AccountTreeNode:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        balance:
          type: number
        children:
          type: array
          items:
            $ref: '#/components/schemas/AccountTreeNode'

    AccountCategory:
      type: string
      enum:
        - short_term_asset
        - long_term_asset
        - short_term_liability
        - long_term_liability
        - equity
        - revenue
        - operating_expense
        - cost_of_goods_sold
        - other_income
        - other_expense
        - manufacturing_cost

    AccountNature:
      type: string
      enum: [debit, credit]

    CreateJournalRequest:
      type: object
      required: [periodId, fiscalYearId, voucherDate, postingDate, description, lines]
      properties:
        periodId:
          type: string
        fiscalYearId:
          type: string
        voucherDate:
          type: string
          format: date
        postingDate:
          type: string
          format: date
        description:
          type: string
        journalType:
          $ref: '#/components/schemas/JournalType'
        voucherTypeId:
          type: string
        voucherSeriesId:
          type: string
        reference:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalLine'
          minItems: 2

    JournalLine:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: string
        debitAmount:
          type: number
          default: 0
        creditAmount:
          type: number
          default: 0
        currencyCode:
          type: string
          default: VND
        description:
          type: string
        costCenterId:
          type: string
        departmentId:
          type: string
        projectId:
          type: string

    JournalType:
      type: string
      enum:
        - standard
        - recurring
        - reversing
        - accrual
        - allocation
        - adjustment
        - closing
        - opening

    JournalEntryStatus:
      type: string
      enum:
        - draft
        - submitted
        - approved
        - posted
        - reversed
        - cancelled

    PostingResult:
      type: object
      properties:
        batchId:
          type: string
        batchNumber:
          type: string
        postedAt:
          type: string
          format: date-time
        totalDebit:
          type: number
        totalCredit:
          type: number
        lineCount:
          type: integer

    CreateFiscalYearRequest:
      type: object
      required: [code, name, startDate, endDate]
      properties:
        code:
          type: string
          example: "FY2026"
        name:
          type: string
          example: "Năm tài chính 2026"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string

  headers:
    X-User-Id:
      description: User ID for audit trail
      type: string
      in: header